requirements.txt
# Core Flask framework and extensions
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-session==0.5.0
flask-cors==4.0.0
werkzeug==2.3.7

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests and API calls
requests==2.31.0

# Image processing and AI
pillow==10.0.1
rembg==2.0.30
openai==1.3.7

# Database (SQLite is built-in with Python)
# No additional database drivers needed for SQLite

# Development and testing (optional)
# pytest==7.4.0
# pytest-flask==1.2.0

# Production deployment (optional)
# gunicorn==21.2.0
# psycopg2-binary==2.9.7  # For PostgreSQL in production